WIP support transmission extension
Note that this does not change our rendering strategy to
TWO_PASSES_ONE_SIDE or TWO_PASSES_TWO_SIDES. According to the
extension spec:

The minimum expectation for a compliant renderer is to render only
opaque objects (i.e. like the first image). Additional transparent
layers may be rendered but are not required.

Ubershader mode does not support this extension when it is combined
with specular-glossiness.  This would exceed our max sampler count
of 16 and increase the shader bloat in gltfio.